requirements.txt
asgiref==3.8.1
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
dj-database-url==2.2.0
Django==5.0.7
django-cors-headers==4.4.0
django-jazzmin==3.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
idna==3.7
numpy==2.0.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
reportlab==4.2.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
